refactor DisplayListChangedEvent to be under events.ui package
